# perf(reporters): overall improvements