@AriPerkkio 2023-03-13
The reason why `renderTree` is slow seems to be these lines:

https://github.com/vitest-dev/vitest/blob/1fe8286c790cef9063b1f5f8147f6d41fc14bf79/packages/vitest/src/node/reporters/renderers/listRenderer.ts#L143-L150

Just by replacing the `output = output.concat(<fn>)` to `output.push(<fn>)` the `renderTree` will reduce from 1200ms to 11ms when called **once** with 5000 tasks. 

![image](https://user-images.githubusercontent.com/14806298/224719583-055552f5-a165-49da-b867-278513694da7.png)

![image](https://user-images.githubusercontent.com/14806298/224719609-51e3776b-6f6e-4822-9190-04536eaa0eb4.png)

In overall the whole test run drops ~1300ms more. Now the total test run of reproduction setup is at 4.63s.