@sheremet-va 2023-03-13
> Just by replacing the output = output.concat(<fn>) to output.push(<fn>) the renderTree will reduce from 1200ms to 11ms when called once with 5000 tasks.

Yes, `concat` is evil!