@AriPerkkio 2023-03-14
`dot` reporter:
- Passing 5000 `c.green('·')` strings to `log-update` is bad for performance. `checkAnsi` + `sliceAnsi` take most of the execution time.
- Passing a single `c.green('·'.repeat(5000))` to `log-update` performs fine

I've refactored dot renderer to group sibling task icons into a single ANSI color code block.